+++ src/__test__/components/GraphVisualisation.test.js
@@ -9,6 +9,7 @@ const setup = () => {
     availableEdges: new DataSet([]),
     availableNodes: new DataSet([]),
     addToArray: jest.fn(),
+    isPhysicsOn: true,
     network: {},
     nodesIdsToDisplay: [],
     physicsEdgeLength: true,